Mark the PyenvLocator tests as functional.
